requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
lxml==5.3.1
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.40
psycopg2==2.9.10
psycopg2-binary==2.9.10
dateparser==1.2.1
schedule==1.2.2